use i16 whereever possible - so we do not have to cast from f32 to i16

+++ src/engine.rs
@@ -26,8 +26,8 @@ impl Image {
         let bounding_box = Rect {
             x: position.x.into(),
             y: position.y.into(),
-            width: element.width() as f32,
-            height: element.height() as f32,
+            width: element.width() as i16,
+            height: element.height() as i16,
         };
         Self {
             element,
@@ -45,17 +45,17 @@ impl Image {
     }
 
     pub fn move_horizontally(&mut self, distance: i16) {
-        self.bounding_box.x += distance as f32;
+        self.bounding_box.x += distance;
         self.position.x += distance;
     }
 
     pub fn set_x(&mut self, x: i16) {
-        self.bounding_box.x = x as f32;
+        self.bounding_box.x = x;
         self.position.x = x;
     }
 
     pub fn right(&self) -> i16 {
-        (self.bounding_box.x + self.bounding_box.width) as i16
+        (self.bounding_box.x + self.bounding_box.width)
     }
 }
 
@@ -140,10 +140,10 @@ pub struct Renderer {
 }
 
 pub struct Rect {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: i16,
+    pub y: i16,
+    pub width: i16,
+    pub height: i16,
 }
 
 impl Rect {
