+++ src/game.rs
@@ -97,19 +97,17 @@ impl RedHatBoy {
     fn destination_box(&self) -> Rect {
         let sprite = self.current_sprite().expect("Cell not found");
         Rect {
-            x: (self.state_machine.context().position.x + sprite.sprite_source_size.x as i16)
-                .into(),
-            y: (self.state_machine.context().position.y + sprite.sprite_source_size.y as i16)
-                .into(),
+            x: (self.state_machine.context().position.x + sprite.sprite_source_size.x).into(),
+            y: (self.state_machine.context().position.y + sprite.sprite_source_size.y).into(),
             width: sprite.frame.w.into(),
             height: sprite.frame.h.into(),
         }
     }
 
     fn bounding_box(&self) -> Rect {
-        const X_OFFSET: f32 = 18.0;
-        const Y_OFFSET: f32 = 14.0;
-        const WIDTH_OFFSET: f32 = 28.0;
+        const X_OFFSET: i16 = 18;
+        const Y_OFFSET: i16 = 14;
+        const WIDTH_OFFSET: i16 = 28;
         let mut bounding_box = self.destination_box();
         bounding_box.x += X_OFFSET;
         bounding_box.width -= WIDTH_OFFSET;
@@ -159,7 +157,7 @@ impl RedHatBoy {
         self.state_machine = self.state_machine.transition(Event::Jump);
     }
 
-    fn land_on(&mut self, y: f32) {
+    fn land_on(&mut self, y: i16) {
         self.state_machine = self.state_machine.transition(Event::Land(y));
     }
 }
@@ -179,7 +177,7 @@ pub enum Event {
     Slide,
     Jump,
     KnockOut,
-    Land(f32),
+    Land(i16),
     Update,
 }
 
@@ -356,8 +354,8 @@ impl Platform {
     }
 
     fn bounding_boxes(&self) -> Vec<Rect> {
-        const X_OFFSET: f32 = 60.0;
-        const END_HEIGHT: f32 = 54.0;
+        const X_OFFSET: i16 = 60;
+        const END_HEIGHT: i16 = 54;
         let destination_box = self.destination_box();
         let bounding_box_one = Rect {
             x: destination_box.x,
@@ -368,7 +366,7 @@ impl Platform {
         let bounding_box_two = Rect {
             x: destination_box.x + X_OFFSET,
             y: destination_box.y,
-            width: destination_box.width - (X_OFFSET * 2.0),
+            width: destination_box.width - (X_OFFSET * 2),
             height: destination_box.height,
         };
         let bounding_box_three = Rect {
@@ -482,9 +480,9 @@ mod red_hat_boy_states {
             }
         }
 
-        pub fn land_on(self, position: f32) -> Self {
+        pub fn land_on(self, position: i16) -> Self {
             RedHatBoyState {
-                context: self.context.set_on(position as i16),
+                context: self.context.set_on(position),
                 _state: Running {},
             }
         }
@@ -519,9 +517,9 @@ mod red_hat_boy_states {
                 _state: Falling {},
             }
         }
-        pub fn land_on(self, position: f32) -> Self {
+        pub fn land_on(self, position: i16) -> Self {
             RedHatBoyState {
-                context: self.context.set_on(position as i16),
+                context: self.context.set_on(position),
                 _state: Sliding {},
             }
         }
@@ -546,9 +544,9 @@ mod red_hat_boy_states {
             JUMPING_FRAME_NAME
         }
 
-        pub fn land_on(self, position: f32) -> RedHatBoyState<Running> {
+        pub fn land_on(self, position: i16) -> RedHatBoyState<Running> {
             RedHatBoyState {
-                context: self.context.reset_frame().set_on(position as i16),
+                context: self.context.reset_frame().set_on(position),
                 _state: Running {},
             }
         }
@@ -596,9 +594,9 @@ mod red_hat_boy_states {
             self
         }
 
-        pub fn land_on(self, position: f32) -> Self {
+        pub fn land_on(self, position: i16) -> Self {
             RedHatBoyState {
-                context: self.context.set_on(position as i16),
+                context: self.context.set_on(position),
                 _state: KnockedOut {},
             }
         }
@@ -698,7 +696,7 @@ impl Game for WalkTheDog {
                         y: HIGH_PLATFORM,
                     },
                 );
-                let background_width = background.width() as i16;
+                let background_width = background.width();
                 Ok(Box::new(WalkTheDog::Loaded(Walk {
                     boy: rhb,
                     backgrounds: [
@@ -706,7 +704,7 @@ impl Game for WalkTheDog {
                         Image::new(
                             background,
                             Point {
-                                x: background_width,
+                                x: background_width as i16,
                                 y: 0,
                             },
                         ),
@@ -781,10 +779,10 @@ impl Game for WalkTheDog {
 
     fn draw(&self, renderer: &Renderer) {
         renderer.clear(&engine::Rect {
-            x: 0.0,
-            y: 0.0,
-            width: 600.0,
-            height: 600.0,
+            x: 0,
+            y: 0,
+            width: 600,
+            height: 600,
         });
         if let WalkTheDog::Loaded(walk) = self {
             walk.backgrounds.iter().for_each(|b| b.draw(renderer));
